Bug fix for timing not found error for mapped trees
This commit fixes and error that was caused by differences in the result of double double addition vs. subtraction in some cases. As a result, the mapped mascot class didn't find the upwards state probabilities for lineages resulting in an error and runs being aborted. This commit changes how the intermediate results of state probbailities on lineages are retrieved by storing the timings of the intermediate results